9393: Remove 'Searchers' namespace from serializers
