+++ app/controllers/responses_controller.rb
@@ -39,7 +39,7 @@ class ResponsesController < ApplicationController
 
         searcher = build_searcher(@responses)
         @responses = apply_searcher_safely(searcher)
-        @searcher_serializer = Searchers::ResponsesSearcherSerializer.new(searcher)
+        @searcher_serializer = ResponsesSearcherSerializer.new(searcher)
 
         @selected_ids = params[:sel]
         @selected_all_pages = params[:select_all_pages]
